Delete analytics module, rename common folder'
