+++ src/Analytics/ToneBar.js
